Doc: A simplified style for rendering docs with QTextBrowser
Add a new documentation template, with simplified CSS rules that
work better when using QTextBrowser as a backend for Qt Assistant
or Qt Creator Help.

Select this new template by default for offline documentation
builds, but keep the old offline template as part of the template
files; use JavaScript to switch to the 'standard' CSS when the
generated files are viewed with a web browser.

Task-number: QTBUG-48322
Change-Id: Ib197896200bb482935f6e9f3a38976133a1e804d
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@digia.com>