qdoc: Fixed linking for overloaded functions
qdoc was failing to create links when a \l command referred
to a function that has an undocumented overload with no
parameters. qdoc would choose that internal function to be
the target, but the link construction would fail.

qdoc now checks the status of the overload that is matched.
If the matching function is marked internal, qdoc keeps
looking for one that is not marked internal.

Change-Id: Iebf296e79dc2554e54f00ef72b6f6c1ba7074f06
Task-number: QTBUG-47991
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>