Add Qt 5.5.1 changelog with updated 5.7 future direction notice
We should carry this in all 5.5.x and 5.6.x releases, to make sure
everyone gets the message. But for this release, we need to note the
changes from what was announced for 5.5.0:

Clang 3.2 cannot compile std::atomic properly, so it will not be
supported. That means the minimum XCode version we'll require will be
5.1, which fortunately does not raise the minimum OS X version that can
compile Qt. See http://llvm.org/bugs/show_bug.cgi?id=12670 for more
information.

ICC 14.x and 15.x on Windows miscompiles pointer-to-members across DLL
boundaries, so signal-slot connections fail. See QTBUG-40281.

Change-Id: I42e7ef1a481840699a8dffff1400a4377fcf4589
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>