Doc: Improve selection of offline template type
Qt 5.6 now includes a version of the offline documentation template
with simplified CSS suited for rendering HTML with a QTextBrowser
backend.

Select the template in qt-html-templates-offline.qdocconf, instead
of the higher-level qt-module-defaults-offline.qdocconf. This is
better because many projects external to qt5 (including Qt Creator)
do not use qt-module-* includes. This way, we can control the
template selection for all projects from a config file.

Change-Id: I766af422d829f3c9519c5a45093473175363d600
Reviewed-by: Martin Smith <martin.smith@digia.com>